Refactor service components
