+++ service/src/main/kotlin/fi/tampere/trevaka/TampereConfig.kt
@@ -5,17 +5,16 @@
 package fi.tampere.trevaka
 
 import fi.espoo.evaka.invoicing.domain.PaymentIntegrationClient
-import fi.espoo.evaka.logging.defaultAccessLoggingValve
 import fi.espoo.evaka.shared.FeatureConfig
 import fi.espoo.evaka.shared.auth.UserRole
 import fi.espoo.evaka.shared.security.actionrule.ActionRuleMapping
 import fi.espoo.evaka.titania.TitaniaEmployeeIdConverter
 import fi.tampere.trevaka.security.TampereActionRuleMapping
-import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory
-import org.springframework.boot.web.server.WebServerFactoryCustomizer
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.core.env.Environment
+import trevaka.titania.TrimStartTitaniaEmployeeIdConverter
+import trevaka.tomcat.tomcatAccessLoggingCustomizer
 
 @Configuration
 class TampereConfig {
@@ -54,13 +53,8 @@ class TampereConfig {
     fun actionRuleMapping(): ActionRuleMapping = TampereActionRuleMapping()
 
     @Bean
-    fun titaniaEmployeeIdConverter(): TitaniaEmployeeIdConverter = object : TitaniaEmployeeIdConverter {
-        override fun fromTitania(employeeId: String): String = employeeId.trimStart('0')
-    }
+    fun titaniaEmployeeIdConverter(): TitaniaEmployeeIdConverter = TrimStartTitaniaEmployeeIdConverter()
 
     @Bean
-    fun tomcatCustomizer(env: Environment) =
-        WebServerFactoryCustomizer<TomcatServletWebServerFactory> {
-            it.addContextValves(defaultAccessLoggingValve(env))
-        }
+    fun accessLoggingCustomizer(env: Environment) = tomcatAccessLoggingCustomizer(env)
 }
